multi_modal_agent/requirements.txt
torch
transformers
Pillow
